Initial Commit
Copied the current version of the mod, with the following features:

New Skills:
Brawl (gained from combat)
Martial Arts (gained from dojo)
Magic (unimplemented)

New Location:
WIP Dojo (take self-defense lessons)

Mechanics:
Animals Disabled

Brawl and Martial Arts do not currently contribute to dodge checks in random events.

Dojo should be functional, but flavor text is not complete. Some text is just copied from Dance Studio.

All animals are beastfolk all the time, including the Black Wolf. The option to change this has been removed from the game. (Because of flavor changes to the world which are forthcoming.)

Minor change to default preference settings, mostly to suit my personal preferences because it's a pain to change everything on a new game every time and it's my mod so I want to cater it to my tastes.