Changed Folder Name
Just changed the folder name to reflect the version number, for the convenience of anyone who downloads multiple versions of the mod.

Current version, I have named 0.1.0a

First digit will be 0 until all major planned features are in the game.

Second digit will increment upwards every time there are moderate changes to the mod, such as a new feature.

Third digit will increment with even the smallest change. But shouldn't go over 9, as minor edits won't warrant a new upload for the most part.

"a" just stands for alpha, will change to "b" for beta once I feel the mod is more complete and more stable. Eventually letters could be removed?

I include the version of DoL that this mod is based off of in addition to the mod's version, just so everything is clear.